requirements.txt
streamlit
langchain
openai
faiss-cpu
tqdm
requests
python-dotenv
